+++ src/inmanta_module_factory/inmanta/implementation.py
@@ -63,6 +63,13 @@ class Implementation(ModuleElement):
             # Entity is in a different file
             entity_path = self.entity.full_path_string
 
+        docstring = self.docstring()
+        if docstring:
+            docstring = f'"""\n{docstring}\n"""\n'
+            docstring = indent(docstring, prefix=INDENT_PREFIX)
+
         return (
-            f"implementation {self.name} for {entity_path}:\n" + indent(self.content.strip(), prefix=INDENT_PREFIX) + "\nend\n"
+            f"implementation {self.name} for {entity_path}:\n{docstring}"
+            + indent(self.content.strip(), prefix=INDENT_PREFIX)
+            + "\nend\n"
         )